edit all sql request

+++ Backend/Invictus_sport_club/src/main/java/com/invictus/services/EsporteService.java
@@ -9,6 +9,7 @@ import com.invictus.domain.esporte.Esporte;
 import com.invictus.domain.esporte.dto.EsporteRequest;
 import com.invictus.mapper.EsporteMapper;
 import com.invictus.repository.EsporteRepository;
+import com.invictus.repository.QuadraRepository;
 import com.invictus.services.esporte.ValidadorEsporteService;
 
 @Service
@@ -23,10 +24,13 @@ public class EsporteService {
 	@Autowired
 	private EsporteMapper esporteMapper;
 
+	@Autowired
+	private QuadraRepository quadraRepository;
+
 	public Esporte cadastrarEsporte(EsporteRequest request) {
 		validadores.forEach((v) -> v.validar(request));
 		Esporte esporte = esporteMapper.esporteRequestToEsporte(request);
-		return esporteRepository.save(esporte);
+		return esporteRepository.saveEsporte(esporte.getDescricao(), esporte.isAtivo());
 	}
 
 	public List<Esporte> listarEsportes() {
@@ -40,6 +44,7 @@ public class EsporteService {
 
 	public Esporte editarEsporte(Long idEsporte, EsporteRequest request) {
 		verificadorEsporteExiste(idEsporte);
+
 		Esporte esporte = esporteMapper.esporteRequestToEsporte(request);
 		esporte.setIdEsporte(idEsporte);
 		return esporteRepository.save(esporte);
@@ -47,6 +52,8 @@ public class EsporteService {
 
 	public void deletarEsporte(Long idEsporte) {
 		verificadorEsporteExiste(idEsporte);
+		if (quadraRepository.existsByEsporteId(idEsporte))
+			throw new RuntimeException("Não é possível excluir esse exporte pois há quadras vinculadas a ele");
 		esporteRepository.deleteById(idEsporte);
 	}
