+++ Backend/Invictus_sport_club/src/main/java/com/invictus/services/TipoQuadraService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.invictus.domain.tipoQuadra.TipoQuadra;
+import com.invictus.repository.QuadraRepository;
 import com.invictus.repository.TipoQuadraRepository;
 
 @Service
@@ -13,6 +14,9 @@ public class TipoQuadraService {
 
 	@Autowired
 	private TipoQuadraRepository tipoQuadraRepository;
+
+	@Autowired
+	private QuadraRepository quadraRepository;
 
 	public TipoQuadra cadastrarTipoQuadra(TipoQuadra tipoQuadra) {
 		return tipoQuadraRepository.save(tipoQuadra);
@@ -35,6 +39,8 @@ public class TipoQuadraService {
 
 	public void deletarTipoQuadraPorId(Long idTipoQuadra) {
 		verificadorTipoQuadraExiste(idTipoQuadra);
+		if (quadraRepository.existsByTipoQuadraId(idTipoQuadra))
+			throw new RuntimeException("Não é possível excluir pois há quadras vinculadas com esse tipo de quadra");
 		tipoQuadraRepository.deleteById(idTipoQuadra);
 	}
